Move commands into subdirectory, create base class
Base class created with potentially common properties from existing class
