+++ Commands/RequireGuildContextAttribute.cs
@@ -1,9 +1,8 @@
 ﻿using Discord.Interactions;
 
-namespace WorldTime;
+namespace WorldTime.Commands;
 /// <summary>
-/// Implements the included precondition from Discord.Net, requiring a guild context while using our custom error message.<br/><br/>
-/// Combining this with <see cref="RequireBotModeratorAttribute"/> is redundant. If possible, only use the latter instead.
+/// Implements the included precondition from Discord.Net, requiring a guild context while using our custom error message.
 /// </summary>
 class RequireGuildContextAttribute : RequireContextAttribute {
     public const string Error = "Command not received within a guild context.";
